third_party/blink/web_tests/http/tests/fetch/chromium/call-extra-crash-construct-and-look.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/stack-overflow.js"></script>
<script>
test(() => {
  let response;
  fillStackAndRun(() => response = new Response('hi'));
  assert_equals(response.body instanceof ReadableStream, true, 'response body should be ReadableStream');
}, 'stack overflow in Response constructor should not crash the browser');
</script>
